refactoring change-password related components
