+++ src/app/components/change-password-page/ChangePasswordContent.tsx
@@ -1,0 +1,12 @@
+import Container from '@mui/material/Container';
+import { ChangePasswordForm } from './ChangePasswordForm';
+
+const ChangePasswordContent = () => {
+  return (
+    <Container maxWidth="sm">
+      <ChangePasswordForm />
+    </Container>
+  );
+};
+
+export default ChangePasswordContent;
