+++ src/app/components/change-password-page/ChangePasswordForm/ChangePasswordFormAlert/index.ts
@@ -1,0 +1,1 @@
+export { default as ChangePasswordFormAlert } from './ChangePasswordFormAlert';
