+++ src/app/components/change-password-page/ChangePasswordForm/ChangePasswordFormBody/index.ts
@@ -1,0 +1,1 @@
+export { default as ChangePasswordFormBody } from './ChangePasswordFormBody';
