+++ src/app/components/change-password-page/ChangePasswordForm/ChangePasswordFormHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as ChangePasswordFormHeader } from './ChangePasswordFormHeader';
