+++ src/app/components/change-password-page/ChangePasswordForm/ChangePasswordFormInputs/index.ts
@@ -1,0 +1,1 @@
+export { default as ChangePasswordFormInputs } from './ChangePasswordFormInputs';
