+++ src/app/components/change-password-page/ChangePasswordForm/ChangePasswordFormSubmit/index.ts
@@ -1,0 +1,1 @@
+export { default as ChangePasswordFormSubmit } from './ChangePasswordFormSubmit';
