+++ src/app/components/change-password-page/index.ts
@@ -1,0 +1,1 @@
+export { default as ChangePasswordContent } from './ChangePasswordContent';
